Remove unecessary and invalid HTML Comments
They were unecessary becaus they just told the reader
where a form started and where it ended, but that's duplicated
information. A form starts with a <form ...> and ends at the
corresponding </form> no need to mark that with comments.

Further Reading:
http://blog.codinghorror.com/code-tells-you-how-comments-tell-you-why/
http://blog.codinghorror.com/coding-without-comments/

They were invalid, because the ending comment lines (<!-------!-->)
contained two (or more) consecutive U+002D HYPHEN-MINUS characters (--)
which is prohibited according to the Spec.

Further Reading:
http://www.w3.org/html/wg/drafts/html/master/syntax.html#comments